views (refactor): cambios por estandarización de las plantillas

+++ views.py
@@ -51,10 +51,6 @@ class QlikContextMixin(PersonalContextMixin):
         context['general']['menu_app'] = apps.get_app_config(__package__).name +'_menu.html'
         return context 
 
-### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### 
-### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### 
-### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### 
-
 class IndexTemplateView(TemplateView, QlikContextMixin):
     template_name = 'template/index.html'
 
@@ -106,31 +102,22 @@ class LicenciasListView(PersonalListView, QlikContextMixin):
     paginate_by = 10
     extra_context = {
         'title': _('Licencias'),
+        'create' :{
+            'url':      TipoLicencia.url_create(),
+            'display':  _('Nuevo'),
+            'img':      'qs_create.png',
+        },
         'campos': {
-            #-1: no enumera
-            # 0: inicia numeración en 0
-            # 1: inicia numeración en 1
             'enumerar': 1,
-            # Si hay valor se muestra opciones por linea, de lo contrario no se muestran
             'opciones': _('Opciones'),
-            # Lista de campos que se deben mostrar en la tabla
             'lista': [
-                'descripcion', 
-                'cantidad', 
+                'descripcion', 'cantidad', 
             ],
         },
         'campos_extra': [
-            {
-                'nombre':   _('Disponibles'), #display
-                # valor, constante o funcion 
-                'funcion': 'licencias_no_asignadas',  
-            },
+            { 'nombre':   _('Disponibles'), 'funcion': 'licencias_no_asignadas' },
         ],
         'opciones': DISPLAYS['opciones'],
-        'create' :{
-            'display':  _('Nuevo'),
-            'url':      TipoLicencia.url_create(),
-        },
         'mensaje': {
             'vacio': DISPLAYS['tabla_vacia'],
         },
@@ -164,34 +151,33 @@ class LicenciasDetailView(PersonalDetailView, QlikContextMixin):
     extra_context = {
         'title': _('Tipo'),
         'campos': {
-            'opciones': _('Opciones'),
             'lista': [
-                #'id',
                 'descripcion', 
                 'cantidad', 
             ],
+            'opciones': _('Opciones'),
         },
+        'campos_extra': [ 
+            {'nombre': _('Licencias asignadas'), 'funcion': 'licencias_asignadas'},
+            {'nombre': _('Licencias pendientes'), 'funcion': 'licencias_no_asignadas'},
+        ],
         'opciones': DISPLAYS['opciones'],
     }
 
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
         context['permisos'] = {
-            'create': self.request.user.has_perm('qliksense.add_tipolicencia'),
+            #'create': self.request.user.has_perm('qliksense.add_tipolicencia'),
             'update': self.request.user.has_perm('qliksense.change_tipolicencia'),
             'delete': self.request.user.has_perm('qliksense.delete_tipolicencia'),
         }
-        context['campos_adicionales'] = [ 
-            {'display': _('Licencias asignadas'), 'valor': self.object.licencias_asignadas()},
-            {'display': _('Licencias pendientes'), 'valor': self.object.licencias_no_asignadas()},
-        ]
         if self.request.user.has_perm('qliksense.add_area_tipolicencia'):
             context['forms'] = [
                 {
                     'modal':    'area_tipolicencia', 
-                    'action':   reverse_lazy('qliksense:create_areatipo')+f'?next='+self.object.url_detail(),
                     'display':  _('Asignación de licencias'),
                     'link_img': 'qs_areatipolicencia.png',
+                    'action':   reverse_lazy('qliksense:create_areatipo')+f'?next='+self.object.url_detail(),
                     'form':     Area_TipoLicencia_ModelForm('tipo', instance=Area_TipoLicencia(tipo=self.object)),
                     'opciones': DISPLAYS['forms'],
                 },
@@ -200,22 +186,14 @@ class LicenciasDetailView(PersonalDetailView, QlikContextMixin):
             context['tables'] = [
                 {
                     'title':        _('Areas'),
-                    'enumerar':     1,
                     'object_list':  Area_TipoLicencia.objects.filter(tipo=self.object).order_by('area__nombre'),
-                    'campos':       ['area', 'cantidad'],
-                    'opciones':     _('Opciones'),
+                    'enumerar':     1,
+                    'lista':       ['area', 'cantidad'],
                     'campos_extra': [
-                        {
-                            'nombre':   _('Asignadas'), #display
-                            # valor, constante o funcion 
-                            'funcion': 'licencias_asignadas',  
-                        },
-                        {
-                            'nombre':   _('Disponibles'), #display
-                            # valor, constante o funcion 
-                            'funcion': 'licencias_disponibles',  
-                        },
+                        { 'nombre':   _('Asignadas'), 'funcion': 'licencias_asignadas' },
+                        { 'nombre':   _('Disponibles'), 'funcion': 'licencias_disponibles' },
                     ],
+                    'opciones':     _('Opciones'),
                     'permisos': {
                         'update':   self.request.user.has_perm('qliksense.change_area_tipolicencia'),
                         'delete':   self.request.user.has_perm('qliksense.delete_area_tipolicencia'),
@@ -257,23 +235,19 @@ class AreaListView(PersonalListView, QlikContextMixin):
     paginate_by = 10
     extra_context = {
         'title': _('Areas'),
+        'create' :{
+            'url':      Area.url_create(),
+            'display':  _('Nueva'),
+            'img':      'qs_create.png',
+        },
         'campos': {
-            #-1: no enumera
-            # 0: inicia numeración en 0
-            # 1: inicia numeración en 1
             'enumerar': 1,
-            # Si hay valor se muestra opciones por linea, de lo contrario no se muestran
-            'opciones': _('Opciones'),
-            # Lista de campos que se deben mostrar en la tabla
             'lista': [
                 'nombre', 
             ],
+            'opciones': _('Opciones'),
         },
         'opciones': DISPLAYS['opciones'],
-        'create' :{
-            'display':  _('Nueva'),
-            'url':      Area.url_create(),
-        },
         'mensaje': {
             'vacio': DISPLAYS['tabla_vacia'],
         },
@@ -310,10 +284,10 @@ class AreaDetailView(PersonalDetailView, QlikContextMixin):
     extra_context = {
         'title': _('Area / Gerencia'),
         'campos': {
-            'opciones': _('Opciones'),
             'lista': [
                 'nombre' 
             ],
+            'opciones': _('Opciones'),
         },
         'opciones': DISPLAYS['opciones'],
     }
@@ -321,7 +295,6 @@ class AreaDetailView(PersonalDetailView, QlikContextMixin):
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
         context['permisos'] = {
-            'create': self.request.user.has_perm('qliksense.add_area'),
             'update': self.request.user.has_perm('qliksense.change_area'),
             'delete': self.request.user.has_perm('qliksense.delete_area'),
         }
@@ -329,9 +302,9 @@ class AreaDetailView(PersonalDetailView, QlikContextMixin):
             context['forms'] = [
                 {
                     'modal':    'area_tipolicencia', 
-                    'action':   reverse_lazy('qliksense:create_areatipo')+f'?next='+self.object.url_detail(),
                     'display':  _('Asignación de licencias'),
                     'link_img': 'qs_areatipolicencia.png',
+                    'action':   reverse_lazy('qliksense:create_areatipo')+f'?next='+self.object.url_detail(),
                     'form':     Area_TipoLicencia_ModelForm('area', instance=Area_TipoLicencia(area=self.object)),
                     'opciones': DISPLAYS['forms'],
                 },
@@ -340,26 +313,18 @@ class AreaDetailView(PersonalDetailView, QlikContextMixin):
             context['tables'] = [
                 {
                     'title':        _('Licencias'),
-                    'enumerar':     1,
                     'object_list':  Area_TipoLicencia.objects.filter(area=self.object).order_by('tipo__descripcion'),
-                    'campos':       ['tipo', 'cantidad'],
+                    'enumerar':     1,
+                    'lista':       ['tipo', 'cantidad'],
                     'campos_extra': [
-                        {
-                            'nombre':   _('Asignadas'), #display
-                            # valor, constante o funcion 
-                            'funcion': 'licencias_asignadas',  
-                        },
-                        {
-                            'nombre':   _('Disponibles'), #display
-                            # valor, constante o funcion 
-                            'funcion': 'licencias_disponibles',  
-                        },
+                        { 'nombre':   _('Asignadas'), 'funcion': 'licencias_asignadas' },
+                        { 'nombre':   _('Disponibles'), 'funcion': 'licencias_disponibles' },
                     ],
+                    'opciones':     _('Opciones'),
                     'permisos': {
                         'update':   self.request.user.has_perm('qliksense.change_area_tipolicencia'),
                         'delete':   self.request.user.has_perm('qliksense.delete_area_tipolicencia'),
                     },
-                    'opciones':     _('Opciones'),
                     #Si tiene next, redirecciona a esa pagina
                     'next':         self.object.url_detail(),
                 },
@@ -463,23 +428,19 @@ class UsuarioListView(PersonalListView, QlikContextMixin):
     paginate_by = 10
     extra_context = {
         'title': _('Usuarios'),
+        'create' :{
+            'url':      Usuario.url_create(),
+            'display':  _('Nuevo'),
+            'img':      'qs_create.png',
+        },
         'campos': {
-            #-1: no enumera
-            # 0: inicia numeración en 0
-            # 1: inicia numeración en 1
             'enumerar': 1,
-            # Si hay valor se muestra opciones por linea, de lo contrario no se muestran
-            'opciones': _('Opciones'),
-            # Lista de campos que se deben mostrar en la tabla
             'lista': [
                 'area_tipo', 'nombre', 'extension', 'correo', 
             ],
+            'opciones': _('Opciones'),
         },
         'opciones': DISPLAYS['opciones'],
-        'create' :{
-            'display':  _('Nuevo'),
-            'url':      Usuario.url_create(),
-        },
         'mensaje': {
             'vacio': DISPLAYS['tabla_vacia'],
         },
@@ -539,30 +500,29 @@ class StreamListView(PersonalListView, QlikContextMixin):
     extra_context = {
         'title': _('Streams'),
         'campos': {
-            #-1: no enumera
-            # 0: inicia numeración en 0
-            # 1: inicia numeración en 1
             'enumerar': 1,
-            # Si hay valor se muestra opciones por linea, de lo contrario no se muestran
-            'opciones': _('Opciones'),
-            # Lista de campos que se deben mostrar en la tabla
             'lista': [
                 'nombre', 
             ],
+            'opciones': _('Opciones'),
         },
         'opciones': DISPLAYS['opciones'],
-        'botones_extra':[
-            {
-                'display':  _('Recargar'),
-                'url':      reverse_lazy('qliksense:refresh_all_data'),
-                'img':      'qs_refresh.png',
-            
-            },
-        ],
         'mensaje': {
             'vacio': DISPLAYS['tabla_vacia'],
         },
     }
+
+    def get_context_data(self):
+        context = super().get_context_data()
+        context['botones_extra'] = [
+            {
+                'permiso':  self.request.user.has_perm('qliksense.add_stream'),
+                'url':      reverse_lazy('qliksense:refresh_all_data'),
+                'display':  _('Recargar'),
+                'img':      'qs_refresh.png',
+            },
+        ],
+        return context
 
 class StreamDetailView(PersonalDetailView, QlikContextMixin):
     permission_required = 'qliksense.view_stream'
@@ -570,12 +530,6 @@ class StreamDetailView(PersonalDetailView, QlikContextMixin):
     model = Stream
     extra_context = {
         'title': _('Stream'),
-        'campos': {
-            #'opciones': _('Opciones'),
-            #'lista': [
-            #    'nombre' 
-            #],
-        },
         'opciones': DISPLAYS['opciones'],
     }
 
@@ -586,12 +540,9 @@ class StreamDetailView(PersonalDetailView, QlikContextMixin):
             context['tables'] = [
                 {
                     'title':        _('Modelos'),
-                    'enumerar':     1,
                     'object_list':  Modelo.objects.filter(stream=self.object).order_by('nombre'),
-                    'campos':       ['nombre'],
-                    'permisos': {
-                        'detail':   self.request.user.has_perm('qliksense.view_modelo'),
-                    },
+                    'enumerar':     1,
+                    'lista':       ['nombre', ],
                     'opciones':     _('Opciones'),
                 },
             ]
@@ -611,52 +562,54 @@ def stream_refresh(request, qs_ws):
         Obtiene los Streams correspondientes, actualiza sus nombres y los crea en caso de ser nuevos.
         Elimina los streams que ya no existen en Qlik.
     '''
-    streams = Stream.objects.all()
-    array_insert = []
+    if request.user.is_authenticated and request.user.has_perm('qliksense.add_stream'):
+        streams = Stream.objects.all()
+        array_insert = []
 
-    for element in qs_ws.get_all_streams():
-        stream = streams.filter(id = element['qId'])
-        streams = streams.exclude(id = element['qId'])
-        if stream:
-            if not stream.filter(nombre = element['qName']):
-                stream[0].nombre = element['qName']
-                stream[0]._history_user = request.user
-                stream[0].save()
-        else:
-            stream = Stream(id = element['qId'], nombre=element['qName'])
-            stream._history_user = request.user
-            array_insert.append(stream)
+        for element in qs_ws.get_all_streams():
+            stream = streams.filter(id = element['qId'])
+            streams = streams.exclude(id = element['qId'])
+            if stream:
+                if not stream.filter(nombre = element['qName']):
+                    stream[0].nombre = element['qName']
+                    stream[0]._history_user = request.user
+                    stream[0].save()
+            else:
+                stream = Stream(id = element['qId'], nombre=element['qName'])
+                stream._history_user = request.user
+                array_insert.append(stream)
 
-    streams.delete()
-    bulk_create_with_history(array_insert, Stream, batch_size=500)
+        streams.delete()
+        bulk_create_with_history(array_insert, Stream, batch_size=500)
 
 def model_refresh(request, qs_ws):
     '''
         Obtiene los Modelos correspondientes, crea y actualiza el modelo de acuerdo con el id generado desde Qlik.
         Elimina modelos que ya no existen en Qlik.
     '''
-    modelos = Modelo.objects.all()
-    array_insert=[]
+    if request.user.is_authenticated and request.user.has_perm('qliksense.add_stream'):
+        modelos = Modelo.objects.all()
+        array_insert=[]
 
-    for element in qs_ws.get_all_models():
-        modelo = modelos.filter(id = element['id'])
-        modelos = modelos.exclude(id = element['id'])
-        if modelo:
-            if not modelo.filter(nombre = element['nombre']) or not modelo.filter(stream_id = element['stream_id']):
-                modelo[0].nombre = element['nombre']
-                modelo[0].stream_id = element['stream_id']
-                modelo[0]._history_user = request.user
-                modelo[0].save()
-        else:
-            modelo = Modelo(id = element['id'], nombre = element['nombre'], stream_id = element['stream_id'])
-            modelo._history_user = request.user
-            array_insert.append(modelo)
-    
-    for m in modelos:
-        m._history_user = request.user
-    modelos.delete()
-    
-    bulk_create_with_history(array_insert, Modelo, batch_size=500)
+        for element in qs_ws.get_all_models():
+            modelo = modelos.filter(id = element['id'])
+            modelos = modelos.exclude(id = element['id'])
+            if modelo:
+                if not modelo.filter(nombre = element['nombre']) or not modelo.filter(stream_id = element['stream_id']):
+                    modelo[0].nombre = element['nombre']
+                    modelo[0].stream_id = element['stream_id']
+                    modelo[0]._history_user = request.user
+                    modelo[0].save()
+            else:
+                modelo = Modelo(id = element['id'], nombre = element['nombre'], stream_id = element['stream_id'])
+                modelo._history_user = request.user
+                array_insert.append(modelo)
+        
+        for m in modelos:
+            m._history_user = request.user
+        modelos.delete()
+        
+        bulk_create_with_history(array_insert, Modelo, batch_size=500)
 
 
 class ModelDetailView(PersonalDetailView, QlikContextMixin):
@@ -665,24 +618,17 @@ class ModelDetailView(PersonalDetailView, QlikContextMixin):
     model = Modelo
     extra_context = {
         'title': _('Modelo'),
-        'campos': {
-            #'opciones': _('Opciones'),
-            #'lista': [
-            #    'nombre' 
-            #],
-        },
         'opciones': DISPLAYS['opciones'],
     }
 
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
-
-        context['manual_tables'] = [
+        context['tables'] = [
             {
                 'title':        _('Campos'),
-                'enumerar':     1,
                 'object_list':  self.campo_valores(self.object),
-                'campos':       ['Nombre', 'Tipo'],
+                'enumerar':     1,
+                'lista':       ['Nombre', 'Tipo'],
                 'opciones':     _('Opciones'),
             },
         ]
